+++ moler/device/state_machine.py
@@ -7,6 +7,7 @@ __email__ = 'michal.ernst@nokia.com'
 import logging
 
 import transitions
+from moler.helpers import ForwardingHandler
 
 forwarding_handler = None
 
@@ -29,29 +30,3 @@ class StateMachine(transitions.Machine):
         if not forwarding_handler:
             forwarding_handler = ForwardingHandler(target_logger_name="moler.state_machine")
             self.logger.addHandler(forwarding_handler)
-
-
-class ForwardingHandler(logging.Handler):
-    """
-    Take log record and pass it to target_logger
-    """
-
-    def __init__(self, target_logger_name):
-        super(ForwardingHandler, self).__init__(level=1)
-        self.target_logger_name = target_logger_name
-        self.target_logger = logging.getLogger('moler')
-
-    def emit(self, record):
-        """
-        Emit a record.
-
-        Output the record to the target_logger, catering for rollover as described
-        in doRollover().
-        """
-        record.name = self.target_logger_name
-
-        if (record.levelno == logging.INFO) or (record.levelname == "INFO"):
-            record.levelno = logging.DEBUG
-            record.levelname = "DEBUG"
-
-        self.target_logger.handle(record)
